utils: move uchroot functions to utils/uchroot
